requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohttp=3.8.3=py311h2bbff1b_0
aiosignal=1.2.0=pyhd3eb1b0_0
appdirs=1.4.4=pyhd3eb1b0_0
async-timeout=4.0.2=py311haa95532_0
attrs=22.1.0=py311haa95532_0
blas=1.0=mkl
bottleneck=1.3.5=py311h5bb9823_0
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
brotlipy=0.7.0=py311h2bbff1b_1002
bzip2=1.0.8=he774522_0
ca-certificates=2023.05.30=haa95532_0
certifi=2023.7.22=py311haa95532_0
cffi=1.15.1=py311h2bbff1b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py311haa95532_0
colorama=0.4.6=py311haa95532_0
contourpy=1.0.5=py311h59b6b97_0
cryptography=41.0.2=py311hac1b9e3_0
cycler=0.11.0=pyhd3eb1b0_0
et_xmlfile=1.1.0=py311haa95532_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=ha860e81_0
frozenlist=1.3.3=py311h2bbff1b_0
giflib=5.2.1=h8cc25b3_3
icc_rt=2022.1.0=h6049295_2
idna=3.4=py311haa95532_0
intel-openmp=2023.1.0=h59b6b97_46319
joblib=1.2.0=py311haa95532_0
jpeg=9e=h2bbff1b_1
kiwisolver=1.4.4=py311hd77b12b_0
lerc=3.0=hd77b12b_0
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libdeflate=1.17=h2bbff1b_0
libffi=3.4.4=hd77b12b_0
libpng=1.6.39=h8cc25b3_0
libtiff=4.5.0=h6c2663c_2
libwebp=1.2.4=hbc33d0d_1
libwebp-base=1.2.4=h2bbff1b_1
lz4-c=1.9.4=h2bbff1b_0
matplotlib-base=3.7.1=py311hf62ec03_1
mkl=2023.1.0=h8bd8f75_46356
mkl-service=2.4.0=py311h2bbff1b_1
mkl_fft=1.3.6=py311hf62ec03_1
mkl_random=1.2.2=py311hf62ec03_1
multidict=6.0.2=py311h2bbff1b_0
munkres=1.1.4=py_0
numexpr=2.8.4=py311h1fcbade_1
numpy=1.25.0=py311hdab7c0b_0
numpy-base=1.25.0=py311hd01c5d8_0
openai=0.27.4=py311haa95532_0
openpyxl=3.0.10=py311h2bbff1b_0
openssl=3.0.9=h2bbff1b_0
packaging=23.0=py311haa95532_0
pandas=1.5.3=py311heda8569_0
pandas-stubs=1.5.3.230203=py311haa95532_0
pillow=9.4.0=py311hd77b12b_0
pip=23.1.2=py311haa95532_0
plotly=5.9.0=py311haa95532_0
pooch=1.4.0=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=23.2.0=py311haa95532_0
pyparsing=3.0.9=py311haa95532_0
pysocks=1.7.1=py311haa95532_0
python=3.11.4=he1021f5_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-dotenv=0.21.0=py311haa95532_0
pytz=2022.7=py311haa95532_0
requests=2.29.0=py311haa95532_0
scikit-learn=1.2.2=py311hd77b12b_1
scipy=1.10.1=py311hc1ccb85_1
setuptools=68.0.0=py311haa95532_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h2bbff1b_0
tbb=2021.8.0=h59b6b97_0
tenacity=8.2.2=py311haa95532_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h2bbff1b_0
tqdm=4.65.0=py311h746a85d_0
types-pytz=2022.4.0.0=py311haa95532_1
tzdata=2023c=h04d1e81_0
urllib3=1.26.16=py311haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.38.4=py311haa95532_0
win_inet_pton=1.1.0=py311haa95532_0
xz=5.4.2=h8cc25b3_0
yarl=1.8.1=py311h2bbff1b_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
